fix: move commands from separate file to existing ones
